requirements.txt
discord.py
python-dotenv
PyNaCl
yt-dlp
psutil
aiohttp
Pillow
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
google-generativeai
discord-py-interactions
ollama
wavelink
beautifulsoup4
chromadb
colorlog
aiofiles